General usability updates
Tested set.
Added new argument to set that allows the array to increase in size if needed.
Added a toMap() method, which allows a Mutable Array++ to be used as a map of Integers to Ts. This replaces the previously-unfinished Map inhritance in Array++.
Improved Array++.toArray().
Added streaming capability to Array++.